# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Documentation",
        "description": "Structure of Iglu's Shop API functions",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/products": {
            "get": {
                "tags": [
                    "Products"
                ],
                "summary": "Get all products without discount",
                "description": "Returns a list of all products without a discount.\n\nThis endpoint is public.\n\nThis function retrieves all products without a discount from the\nProduct model and returns them in JSON format. Each product includes\nid, name, description, price, stock, image, has_discount, discount, provider_id, category_id, created_at, and updated_at.",
                "operationId": "d97d1c3727cbac49708b85f7d11f7d58",
                "responses": {
                    "200": {
                        "description": "List of products without discount (public endpoint, does not require authentication)",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Product"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Products"
                ],
                "summary": "Create a new product",
                "description": "Creates a new product with the provided data.\n\nRequires authentication and administrator permissions.\n\nWhen creating a new product, the following fields are required:\n\n- `name`: The name of the product. Must be a string with a minimum of 3 characters.\n- `description`: A detailed description of the product. Must be a string with a maximum of 255 characters.\n- `price`: The price of the product. Must be a decimal number with up to two decimal places.\n- `category_id`: The ID of the category the product belongs to. Must be an integer and must exist in the database.\n- `provider_id`: The ID of the provider associated with the product. Must be an integer and must exist in the database.\n\nThe following fields are optional, but if included, must meet the specified validations:\n\n- `stock`: An integer indicating the available quantity of the product in stock.\n- `image`: A valid URL pointing to the product's image.\n- `has_discount`: A boolean field indicating whether the product has a discount. This field must always be false for standard products and true for outlet products.\n- `discount`: The discount applied to the product. If `has_discount` is false, this value must be 0.\n\nValidation includes type constraints and specific value rules. If any field does not comply with the rules, an error will be returned.",
                "operationId": "785644b7d7dd38af4b51dafd0fbc1b6d",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ProductCreate"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Product successfully created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Product successfully created."
                                        },
                                        "id": {
                                            "type": "integer",
                                            "example": 21
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Unauthorized"
                    },
                    "422": {
                        "description": "Validation errors",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ValidationError"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/products/{id}": {
            "get": {
                "tags": [
                    "Products"
                ],
                "summary": "Get a product by ID",
                "description": "REtrieves data about an especific product.\n\nThis is a public endpoint \n\nBusca el producto por su ID y comprueba que no pertenezca al outlet.\nSi no se encuentra, devuelve un error",
                "operationId": "c87fb2e4d617e7b36c712cad358acdb8",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Product ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Product details",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Product"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Product not found or belonging to outlet",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/NotFoundError"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Products"
                ],
                "summary": "Update a product",
                "description": "Updates the data of an existing product, whether it is from the outlet or not.\n\nRequires authentication and administrator permissions.  \n\nAllows updating one or more fields of the product. \nReturns the updated product if everything is correct. \n\nAllows updating the `has_discount` field to a truthy value in order to move the product\nto the outlet endpoint.",
                "operationId": "416e4959a3919547eb609d87e3db9135",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the product to update",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ProductUpdate"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Product successfully updated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Product successfully updated"
                                        },
                                        "product": {
                                            "$ref": "#/components/schemas/ProductUpdate"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Product not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/NotFoundError"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation errors",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ValidationError"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Products"
                ],
                "summary": "Deletes a product",
                "description": "Elimina un producto que no pertenezca al outlet por su ID. \n\nRequiere autenticación y permisos de administrador.\n\nSi el producto no existe, devuelve un error 404. \nSi se elimina correctamente, devuelve un código 204.",
                "operationId": "b49ce9f3cb168b233245548fb4ef36d3",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the product to delete",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Product successfully deleted"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Product not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/NotFoundError"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "NotFoundError": {
                "title": "Resource not found",
                "properties": {
                    "message": {
                        "type": "string",
                        "example": "Resource not found"
                    }
                },
                "type": "object"
            },
...
            "ProductCreate": {
                "required": [
                    "name",
                    "description",
                    "price",
                    "category_id",
                    "provider_id"
                ],
                "properties": {
                    "name": {
                        "type": "string",
                        "minLength": 3,
                        "example": "Sports T-shirt"
                    },
                    "description": {
                        "type": "string",
                        "maxLength": 255,
                        "example": "High-performance technical t-shirt"
                    },
                    "price": {
                        "type": "number",
                        "format": "float",
                        "example": 29.99
                    },
                    "stock": {
                        "type": "integer",
                        "example": 100
                    },
                    "image": {
                        "type": "string",
                        "format": "uri",
                        "example": "https://example.com/product.jpg"
                    },
                    "has_discount": {
                        "description": "Must be false or equivalent (declined). The created product must NOT have a discount initially.",
                        "type": "boolean",
                        "enum": [
                            false
                        ],
                        "example": false
                    },
                    "discount": {
                        "description": "Must be 0 if there is no discount.",
                        "type": "number",
                        "example": 0
                    },
                    "category_id": {
                        "description": "Related category ID (foreign key to Category)",
                        "type": "integer",
                        "example": 1
                    },
                    "provider_id": {
                        "description": "Related provider ID (foreign key to Provider)",
                        "type": "integer",
                        "example": 5
                    }
                },
                "type": "object"
            },
            "Product": {
                "title": "Product",
                "required": [
                    "id",
                    "name",
                    "description",
                    "price",
                    "stock",
                    "image",
                    "has_discount",
                    "discount",
                    "provider_id",
                    "category_id",
                    "created_at",
                    "updated_at"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "Sports T-shirt"
                    },
                    "description": {
                        "type": "string",
                        "example": "High-performance technical t-shirt"
                    },
                    "price": {
                        "type": "number",
                        "example": 29.99
                    },
                    "stock": {
                        "type": "integer",
                        "example": 2
                    },
                    "image": {
                        "type": "string",
                        "format": "uri",
                        "example": "https://www.example.com/images/tshirt.jpg"
                    },
                    "has_discount": {
                        "description": "Este campo debe ser siempre false. si se cambia a true el producto pasa al outlet",
                        "type": "boolean",
                        "default": 0,
                        "example": false
                    },
                    "discount": {
                        "description": "Si has_discount es falsy entonces discount debe ser 0",
                        "type": "number",
                        "example": 0
                    },
                    "provider_id": {
                        "description": "Related provider ID (foreign key a Provider)",
                        "type": "integer",
                        "example": 1
                    },
                    "category_id": {
                        "description": "Related category ID (foreign key a Category)",
                        "type": "integer",
                        "example": 1
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-04-05 21:45:59"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-04-05 21:45:59"
                    }
                },
                "type": "object"
            },
            "ProductUpdate": {
                "title": "Actualización de Producto",
                "description": "Schema para actualizar un producto",
                "properties": {
                    "name": {
                        "type": "string",
                        "minLength": 3,
                        "example": "Sports T-shirt"
                    },
                    "description": {
                        "type": "string",
                        "maxLength": 255,
                        "example": "confortable t-shirt for training"
                    },
                    "price": {
                        "type": "number",
                        "format": "float",
                        "example": 49.99
                    },
                    "stock": {
                        "type": "integer",
                        "example": 150
                    },
                    "image": {
                        "type": "string",
                        "format": "uri",
                        "example": "https://example.com/images/product.png"
                    },
                    "has_discount": {
                        "type": "boolean",
                        "example": false
                    },
                    "discount": {
                        "type": "number",
                        "example": 0
                    },
                    "category_id": {
                        "description": "Related category ID (foreign key a Category)",
                        "type": "integer",
                        "example": 1
                    },
                    "provider_id": {
                        "description": "Related provider ID (foreign key a Provider)",
                        "type": "integer",
                        "example": 2
                    }
                },
                "type": "object"
            },
...
            "ValidationError": {
                "title": "Error de validación",
                "properties": {
                    "message": {
                        "type": "string",
                        "example": "The given data was invalid."
                    },
                    "errors": {
                        "properties": {
                            "error": {
                                "type": "array",
                                "items": {
                                    "type": "string",
                                    "example": "Name is required"
                                }
                            }
                        },
                        "type": "object"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}